Separated clipper server and clipper
